MDTZ-1042: Refactor auth middlewares
