+++ src/web/middleware/auth/connect-server-symmetric-jwt-auth-middleware.ts
