+++ src/web/middleware/auth/index.ts
@@ -1,1 +1,2 @@
 export * from './connect-context-symmetric-jwt-auth-middleware';
+export * from './connect-server-symmetric-jwt-auth-middleware';
