+++ src/web/middleware/index.ts
@@ -1,4 +1,4 @@
 export * from './http-logger-middleware';
-export * from './auth/auth-header-jwt-middleware';
+export * from './auth/connect-server-symmetric-jwt-auth-middleware';
 export * from './error-handler-middleware';
 export * from './extract-user-id-from-headers-middleware';
